test cases; improve readability
